[Refactor] seperated apiviews to each view
